Add in some more tests for the coordinators.
Mostly the stage decider does the actual logic, but it doesn't hurt to
test that the coordinator is using the stage decider properly.
